removed companies in my route handlers, check for ID in put request and 200 response for successful put request
